requirements.txt
PyQt5==5.15.9
python-barcode==0.15.1
qrcode[pil]==7.4.2
Pillow==9.5.0
pandas==2.0.3
python-dateutil==2.8.2
numpy==1.24.3